manifest: introduce internal/manifest.ListInternal and freeze manifest.List
Flips dependency of internal and public manifest.List API and introduces
private manifest.List API which will be extended in future for new
features.

[NO NEW TESTS NEEDED]
[NO TESTS NEEDED]
