+++ internal/manifest/testdata/schema2-to-schema1-by-docker.json
@@ -1,0 +1,1 @@
+../../../internal/image/fixtures/schema2-to-schema1-by-docker.json
