+++ internal/manifest/testdata/unknown-version.manifest.json
@@ -1,0 +1,5 @@
+{
+   "schemaVersion": 99999,
+   "name": "mitr/noversion-nonsense",
+   "tag": "latest"
+}
