+++ manifest/common.go
@@ -1,79 +1,12 @@
 package manifest
 
 import (
-	"encoding/json"
 	"fmt"
 
 	compressiontypes "github.com/containers/image/v5/pkg/compression/types"
 	"github.com/containers/image/v5/types"
 	"github.com/sirupsen/logrus"
 )
-
-// allowedManifestFields is a bit mask of “essential” manifest fields that validateUnambiguousManifestFormat
-// can expect to be present.
-type allowedManifestFields int
-
-const (
-	allowedFieldConfig allowedManifestFields = 1 << iota
-	allowedFieldFSLayers
-	allowedFieldHistory
-	allowedFieldLayers
-	allowedFieldManifests
-	allowedFieldFirstUnusedBit // Keep this at the end!
-)
-
-// validateUnambiguousManifestFormat rejects manifests (incl. multi-arch) that look like more than
-// one kind we currently recognize, i.e. if they contain any of the known “essential” format fields
-// other than the ones the caller specifically allows.
-// expectedMIMEType is used only for diagnostics.
-// NOTE: The caller should do the non-heuristic validations (e.g. check for any specified format
-// identification/version, or other “magic numbers”) before calling this, to cleanly reject unambiguous
-// data that just isn’t what was expected, as opposed to actually ambiguous data.
-func validateUnambiguousManifestFormat(manifest []byte, expectedMIMEType string,
-	allowed allowedManifestFields) error {
-	if allowed >= allowedFieldFirstUnusedBit {
-		return fmt.Errorf("internal error: invalid allowedManifestFields value %#v", allowed)
-	}
-	// Use a private type to decode, not just a map[string]any, because we want
-	// to also reject case-insensitive matches (which would be used by Go when really decoding
-	// the manifest).
-	// (It is expected that as manifest formats are added or extended over time, more fields will be added
-	// here.)
-	detectedFields := struct {
-		Config    any `json:"config"`
-		FSLayers  any `json:"fsLayers"`
-		History   any `json:"history"`
-		Layers    any `json:"layers"`
-		Manifests any `json:"manifests"`
-	}{}
-	if err := json.Unmarshal(manifest, &detectedFields); err != nil {
-		// The caller was supposed to already validate version numbers, so this should not happen;
-		// let’s not bother with making this error “nice”.
-		return err
-	}
-	unexpected := []string{}
-	// Sadly this isn’t easy to automate in Go, without reflection. So, copy&paste.
-	if detectedFields.Config != nil && (allowed&allowedFieldConfig) == 0 {
-		unexpected = append(unexpected, "config")
-	}
-	if detectedFields.FSLayers != nil && (allowed&allowedFieldFSLayers) == 0 {
-		unexpected = append(unexpected, "fsLayers")
-	}
-	if detectedFields.History != nil && (allowed&allowedFieldHistory) == 0 {
-		unexpected = append(unexpected, "history")
-	}
-	if detectedFields.Layers != nil && (allowed&allowedFieldLayers) == 0 {
-		unexpected = append(unexpected, "layers")
-	}
-	if detectedFields.Manifests != nil && (allowed&allowedFieldManifests) == 0 {
-		unexpected = append(unexpected, "manifests")
-	}
-	if len(unexpected) != 0 {
-		return fmt.Errorf(`rejecting ambiguous manifest, unexpected fields %#v in supposedly %s`,
-			unexpected, expectedMIMEType)
-	}
-	return nil
-}
 
 // layerInfosToStrings converts a list of layer infos, presumably obtained from a Manifest.LayerInfos()
 // method call, into a format suitable for inclusion in a types.ImageInspectInfo structure.
