+++ manifest/common_test.go
@@ -15,56 +15,6 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestValidateUnambiguousManifestFormat(t *testing.T) {
-	const allAllowedFields = allowedFieldFirstUnusedBit - 1
-	const mt = "text/plain" // Just some MIME type that shows up in error messages
-
-	type test struct {
-		manifest string
-		allowed  allowedManifestFields
-	}
-
-	// Smoke tests: Success
-	for _, c := range []test{
-		{"{}", allAllowedFields},
-		{"{}", 0},
-	} {
-		err := validateUnambiguousManifestFormat([]byte(c.manifest), mt, c.allowed)
-		assert.NoError(t, err, c)
-	}
-	// Smoke tests: Failure
-	for _, c := range []test{
-		{"{}", allowedFieldFirstUnusedBit}, // Invalid "allowed"
-		{"@", allAllowedFields},            // Invalid JSON
-	} {
-		err := validateUnambiguousManifestFormat([]byte(c.manifest), mt, c.allowed)
-		assert.Error(t, err, c)
-	}
-
-	fields := map[allowedManifestFields]string{
-		allowedFieldConfig:    "config",
-		allowedFieldFSLayers:  "fsLayers",
-		allowedFieldHistory:   "history",
-		allowedFieldLayers:    "layers",
-		allowedFieldManifests: "manifests",
-	}
-	// Ensure this test covers all defined allowedManifestFields values
-	allFields := allowedManifestFields(0)
-	for k := range fields {
-		allFields |= k
-	}
-	assert.Equal(t, allAllowedFields, allFields)
-
-	// Every single field is allowed by its bit, and rejected by any other bit
-	for bit, fieldName := range fields {
-		json := []byte(fmt.Sprintf(`{"%s":[]}`, fieldName))
-		err := validateUnambiguousManifestFormat(json, mt, bit)
-		assert.NoError(t, err, fieldName)
-		err = validateUnambiguousManifestFormat(json, mt, allAllowedFields^bit)
-		assert.Error(t, err, fieldName)
-	}
-}
-
 // Test that parser() rejects all of the provided manifest fixtures.
 // Intended to help test manifest parsers' detection of schema mismatches.
 func testManifestFixturesAreRejected(t *testing.T, parser func([]byte) error, fixtures []string) {
