+++ manifest/fixtures/ociv1.artifact.json
@@ -1,10 +1,1 @@
-{
-  "schemaVersion": 2,
-  "mediaType": "application/vnd.oci.image.manifest.v1+json",
-  "config": {
-    "mediaType": "application/vnd.oci.custom.artifact.config.v1+json",
-    "digest": "",
-    "size": 0
-  },
-  "layers": null
-}
+../../internal/manifest/testdata/ociv1.artifact.json
