+++ manifest/fixtures/ociv1nomime.artifact.json
@@ -1,9 +1,1 @@
-{
-  "schemaVersion": 2,
-  "config": {
-    "mediaType": "application/vnd.oci.custom.artifact.config.v1+json",
-    "digest": "",
-    "size": 0
-  },
-  "layers": null
-}
+../../internal/manifest/testdata/ociv1nomime.artifact.json
