+++ manifest/fixtures/schema2-to-schema1-by-docker.json
@@ -1,1 +1,1 @@
-../../internal/image/fixtures/schema2-to-schema1-by-docker.json
+../../internal/manifest/testdata/schema2-to-schema1-by-docker.json
