+++ manifest/fixtures/v2s2nomime.manifest.json
@@ -1,10 +1,1 @@
-{
-    "schemaVersion": 2,
-    "config": {
-        "mediaType": "application/vnd.docker.container.image.v1+json",
-        "size": 7023,
-        "digest": "sha256:b5b2b2c507a0944348e0303114d8d93aaaa081732b86451d9bce1f432a537bc7"
-    },
-    "layers": [
-    ]
-}
+../../internal/manifest/testdata/v2s2nomime.manifest.json
